backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Werkzeug==3.0.1
SQLAlchemy==2.0.36
PyMySQL==1.1.0
cryptography==41.0.7
greenlet==3.1.1
python-dotenv==1.0.0
google-generativeai
elevenlabs
Pillow
gunicorn
mysql-connector-python
google.genai